update done task4 pr 11

+++ src/ru/mirea/canh/pr11/task3/Const.java
@@ -1,9 +1,9 @@
 package ru.mirea.canh.pr11.task3;
 
 public class Const implements Input {
-    final double constant;
+    final int constant;
 
-    public Const(double constant) {
+    public Const(int constant) {
         this.constant = constant;
     }
 }
