+++ src/ru/mirea/canh/pr11/task4/WhileSpaceCharacters.java
@@ -5,9 +5,7 @@ public class WhileSpaceCharacters implements TripleExpression {
     }
 
     @Override
-    public double evaluate(double x, double y, double z) {
-        //một số ký tự khoảng trắng tùy ý ở bất kỳ đâu (nhưng không phải bên trong hằng số).
-        // need cre exception
+    public int evaluate(int x, int y, int z) {
         return 0;
     }
 }
